Edited (the correct manuscript this time)
The original length of it (without counting the first page and the references section, but counting tables and figures) was 7,525. I managed to lower it by 1056 words, to 6469 words, but ran out of time. I edited sections 1-3, the conclusions section, and the first paragraph of section 4, but wasn't able to look at Sections 5 and 6.